requirements.txt
streamlit
pandas
numpy
scipy
matplotlib
plotly
jinja2
unidecode
openpyxl
xlsxwriter
python-dotenv
google-generativeai
openai
fastapi
uvicorn
python-multipart
requests
reportlab
babel
websockets>=11.0
redis>=5.0
celery>=5.3
aiohttp>=3.8.0
psutil>=5.9.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
openpyxl>=3.1.0
Pillow>=10.0.0
nest-asyncio>=1.5.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
